requirements.txt
streamlit
supabase
python-dotenv
sounddevice
wavio